COMMERCE-1236 Moved dto converter interfaces into headless commerce core
